// api/src/typings/body/ServiceBody.ts
/* eslint-disable @typescript-eslint/no-explicit-any */
import { DiscountTypes } from '../../database/models/Service';
import Job from '../../database/models/Job';
import ServiceItem from '../../database/models/ServiceItem';
import Schedule from '../../database/models/Schedule';
import ChecklistJob from '../../database/models/ChecklistJob';
import JobLabel from '../../database/models/JobLabel';
import { ContactPersonResponseModel, Select } from '../ResponseFormats';
import { ClientBody } from './ClientBody';

export interface ServiceBody {
  id: number;
  serviceType?: string;
  serviceNumber: string;
  serviceTitle?: string;
  description?: string;
  serviceStatus?: string;
  needGST?: boolean;
  termStart?: Date;
  termEnd?: Date;
  contractAmount?: number;
  contractDiscount?: number;
  totalAmount?: number;
  gstAmount?: number;
  grandTotal?: number;
  gstTax?: number;
  salesPerson?: string;
  skills?: any[];
  serviceAddress?: string;
  postalCode?: string;
  billingAddress?: string;
  billingPostalCode?: string;
  clientName?: string;
  entityName?: string;
  clientType?: string;
  clientAgent?: string;
  clientRemarks?: string;
  Jobs: Job[] | any[];
  invoiceId?: number;
  invoiceNumber?: string;
  invoiceStatus?: string;
  discountType?: DiscountTypes;
  discountAmount?: number;
  remarks?: string;
  termCondition?: string;
  clientId?: number;
  ClientBody?: ClientBody;
  serviceAddressId?: number;
  entityId?: number;
  jobId?: number;
  ServiceItems?: ServiceItem[];
  Schedules?: Schedule[];
  Schedule?: Schedule;
  Checklists?: ChecklistJob[];
  JobLabels?: JobLabel[];
  isNextDay?: boolean;
  holidaysDate?: any[];
  totalJob?: number;
  isNewGenerated?: boolean;
  CustomFields?: any[];
  issueDate?: Date;
  expiryDate?: Date;
  ContactPersons?: ContactPersonResponseModel[];
  serviceTemplateId?: number;
  isSendReportToClientEmail?: boolean;
  contractSign?: string;
  selectedEmployees?: Select[];
  selectedVehicles?: Select[];
  isAssignFirstJob?: boolean;
}